// apps/web-svelte/src/lib/api/generated/model/playbackStatisticsBucket.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface PlaybackStatisticsBucket {
  date: string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  accessedCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  completedCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)$ */
  skippedCount: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$ */
  activeSeconds: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$ */
  viewingSeconds: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$ */
  listeningSeconds: number | string;
  /** @pattern ^-?(?:0|[1-9]\d*)(?:\.\d+)?(?:[eE][+-]?\d+)?$ */
  readingSeconds: number | string;
}